# spiffworkflow-backend/src/spiffworkflow_backend/api.yml
openapi: "3.0.2"
info:
  version: 1.0.0
  title: spiffworkflow-backend
  license:
    name: LGPL
servers:
  - url: http://167.86.89.45:5000/v1.0
# this is handled in flask now
security: []
paths:
...
  /login_return:
    parameters:
      - name: code
        in: query
        required: false
        schema:
          type: string
      - name: state
        in: query
        required: true
        schema:
          type: string
      - name: session_state
        in: query
        required: false
        schema:
          type: string
      - name: error
        in: query
        required: false
        schema:
          type: string
      - name: error_description
        in: query
        required: false
        schema:
          type: string
    get:
      operationId: spiffworkflow_backend.routes.authentication_controller.login_return
      tags:
        - Authentication
      responses:
        "200":
          description: Test Return Response
...
  /login_with_access_token:
    parameters:
      - name: access_token
        in: query
        required: true
        schema:
          type: string
      - name: authentication_identifier
        in: query
        required: true
        schema:
          type: string
    post:
      operationId: spiffworkflow_backend.routes.authentication_controller.login_with_access_token
      summary: Authenticate user for API access with an openid token already posessed.
      tags:
        - Authentication
      responses:
        "200":
          description: "Returns ok: true if user successfully logged in."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkTrue"
...
components:
  schemas:
...
    OkTrue:
      properties:
        ok:
          type: boolean
          example: true
...